requirements.txt
flask
beautifulsoup4
newspaper3k
trafilatura